fix: resolves button as shared by the project
